requirements.txt
fastapi==0.136.1
uvicorn==0.46.0
openai==2.36.0
python-dotenv==1.2.2
httpx==0.28.1
